+++ docarray/array/storage/base/getsetdel.py
@@ -324,7 +324,3 @@ class BaseGetSetDelMixin(ABC):
     @abstractmethod
     def _save_offset2ids(self):
         ...
-
-    def sync(self):
-        if hasattr(self, '_offset2ids'):
-            self._save_offset2ids()
